Re-organized the calculators to better group approaches together, updated README.md accordingly
