+++ src/primes/index.ts
@@ -1,7 +1,4 @@
-import linearSieve from './calculators/linear-sieve'
-import naiveCalculator from './calculators/naive'
-import optimizedNaive from './calculators/optimized-naive'
-import sieveOfEratosthenes from './calculators/sieve'
+import { naiveApproach, sieveApproach } from './calculators/calculators'
 import { Comparison } from './types'
 import { runComparisons } from './utils/comparisons'
 
@@ -9,10 +6,10 @@ const max = 1128000
 const min = 0
 
 const comparisonArr: Comparison[] = [
-  ['Naive approach', naiveCalculator],
-  ['Optimized naive approach', optimizedNaive],
-  ['Basic sieve of Eratosthenes approach', sieveOfEratosthenes],
-  ['Linear sieve of Eratosthenes approach', linearSieve],
+  ['Naive approach', naiveApproach.basic],
+  ['Optimized naive approach', naiveApproach.optimized],
+  ['Basic sieve of Eratosthenes approach', sieveApproach.basic],
+  ['Linear sieve of Eratosthenes approach', sieveApproach.linear],
 ]
 
 console.log('--------------')
